requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
pandas==2.1.4
